Switch to rust-vmm's vm-virtio crate
Switch to using rust-vmm's vm-virtio crate and the independent version
of the vhost-user-backend crate.

This breaks some of the tests, as rust-vmm's vm-virtio::Queue doesn't
provide a public method for instantiating a DescriptorChain. There are
plans in vm-virtio to separate the test auxiliary code and make it
reusage [1], so let's allow this breakage now with the intention of
fixing it in the near future.

[1] https://github.com/rust-vmm/vm-virtio/issues/23

Signed-off-by: Sergio Lopez <slp@redhat.com>