Store the InodeAltKey in its respective InodeData
To detect submount points, we are going to need the parent's st_dev when
looking up a node.  An easy way to do this is to store it as part of the
InodeData; and why not store the whole InodeAltKey?  After all, the C
virtiofsd does the same, and we already store the inode ID (which is the
primary ID), too.
(Admittedly, storing the whole InodeAltKey object will only really make
make sense once it contains the mount ID in addition to st_dev and
st_ino (which it should), but we will need statx() support for that.)

For now, this value is dead code, but that will change soon.

Signed-off-by: Max Reitz <mreitz@redhat.com>