Add passthrough::PassthroughFs.announce_submounts
This field will determine whether we announce submounts to the client;
it depends on whether it was enabled in the configuration, and whether
the client supports submounts.  (If the client does not support them but
announcing submounts was enabled in the configuration, warn on stderr.)

Contrary to Config.announce_submounts (which is true by default), this
field is initially false, because we need to check whether the client
supports submounts before setting it to true.

Signed-off-by: Max Reitz <mreitz@redhat.com>