Announce submounts
When configured accordingly (i.e., when not disabled via
--no-announce-submounts) and when the client supports it, let
PassthroughFs::do_lookup() check for every directory whether its st_dev
differs from its parent, and if so, announce this entry to be a submount
point.

Note that it is possible to mount a device inside of an existing mount
of itself, so there can be submounts where st_dev of the mount point
does not differ from the parent's st_dev.  However, in such cases,
virtiofsd-rs will often be utterly confused anyway, because it is very
likely that this leads to collision of the st_dev/st_ino combination we
use as a key to identify nodes.  To get around this, we will need to use
the mount ID as an additional key (which we can get with statx()), but
all in all that is a problem largely unrelated to the submount feature.

Signed-off-by: Max Reitz <mreitz@redhat.com>