Add --no-announce-submounts switch
Providing this switch will clear the announce_submounts field in the
configuration.  (It is set by default.)

(The name "no-announce-submounts" might be a bit weird, but it best
describes the actual behavior: Something like "hide-submounts" might
imply that submounts are not shown to the guest at all; however, they
are shown, they just are not announced as being submounts.  Also,
"no-announce-submounts" pairs very obviously with "announce-submounts"
from C virtiofsd, and also allows for grepping the latter string in
virtiofsd-rs's help text.)

Note that this switch right now does not exactly do what you would think
it would: Because the functionality to announce submounts is not yet
implemented, passing or not passing it makes no difference.  However,
should users want to inhibit submount functionality regardless of the
exact virtiofsd-rs version (e.g. when bisecting), it must be possible to
disable announcing submounts before that is implemented.

Signed-off-by: Max Reitz <mreitz@redhat.com>