Acknowledge the SUBMOUNTS capability
The most basic support for the FUSE SUBMOUNTS capability is to just echo
it back to the kernel, without using the Attr.flags field it implies.
(Attr.flags is still Attr.padding in virtiofsd-rs for now, but as long
as we always set it to 0 (which we do by virtue of the derived Default
implementation), this use is safe.)

Future patches will let rename Attr.padding to Attr.flags and let our
file system code set the SUBMOUNT flag there, so we can make actual use
of this capability.

Signed-off-by: Max Reitz <mreitz@redhat.com>