Add flags field to fuse::Attr
The new flags field replaces the existing padding.

Because those flags cannot be inferred from the libc::stat64 structure,
constructing a fuse::Attr object now requires an additional argument,
namely those flags.  Add a new constructor Attr::with_flags() to do so.

(Alternatively, we could implement From<(libc::stat64, u32)>, but that
is rather unintuitive, so I would rather not do that.)

Along with the field, add the only flag currently allowed, which is the
ATTR_SUBMOUNT flag.  At this point it is still dead code, but that will
change soon.

Signed-off-by: Max Reitz <mreitz@redhat.com>