Add mount ID to stat() result
Encapsulate the stat() result in a new struct and add the mount ID to
it.  For our current stat() (which calls fstatat64()), this ID is always
0 (i.e., unknown), but statx() (which a later patch is going to add)
will set it.

(Also, this mount ID is never read yet.  Other later patches will change
that, too.)

Note that some stat() callers now take the form stat()?.st to change the
following code as little as possible, whereas others now store the new
StatResult object.  These other places are the ones where we are going
to make use of the mnt_id field.

Signed-off-by: Max Reitz <mreitz@redhat.com>