Add statx() as the preferred stat() alternative
The advantage of statx() over fstatat64() (and other stat variants) is
that it returns the mount ID, which allows distinguishing between
different mounts of the same device (which is why a later patch will add
the mount ID to InodeAltKey).

Older kernels may not support statx() yet, so we should check for its
presence at startup (in PassthroughFs::new()) and disable its use if it
is not present (indicated by an ENOSYS error).  We should only do that
test once, and not switch to stat64() if we later receive another
ENOSYS, because once we use the mount ID as part of the inode key,
switching to stat64() would mean to probably lose access to the inodes
indexed so far (with a non-zero mount ID).

Signed-off-by: Max Reitz <mreitz@redhat.com>