Add mount ID to the InodeAltKey
This allows us to distinguish between nodes that are on the same device
(same dev_t), but in different mounts (different mount ID).  That can be
important, because two mounts of the same device can be different (e.g.
one may contain a submount, the other does not), so this is a bug fix
for such cases.

Signed-off-by: Max Reitz <mreitz@redhat.com>