Implement support for multiple sandboxing modes
The daemon may be run in environments where is not possible to create
a sandbox by using Linux namespaces, so add also support for creating
a sandbox using "chroot".

Instead of adding a new flag, rename the boolean "disable-sandbox"
argument to a "sandbox" one with multiple values: "namespace",
"chroot" and "none".

Signed-off-by: Sergio Lopez <slp@redhat.com>