Add initial xattr map parsing
The C-flavored virtiofs daemon allows users to pass a string
containing a series of rules which effectively form a translation
layer for extended attributes between the host and the guest.

This commit adds support for parsing these mapping rules into a
data structure which will be used in a later patch to intercept
{get,set,remove,list}xattr mechanisms to apply the rules specified
in the mapping.

A more detailed summary of xattr mapping rules can be found here:

https://gitlab.com/qemu-project/qemu/-/blob/master/docs/tools/virtiofsd.rst

Signed-off-by: Connor Kuehl <ckuehl@redhat.com>