Add xattr map rule application mechanisms
There are two "directions" in which an xattr mapping rule
is applied: server and client. The direction corresponds
to the source whose data will pass through the xattr mapping
rules. This is significant because the direction determines
how a rule is applied since a rule can apply to one or both
directions and the actions it takes depends on the direction
it is applied in.

For example, the {get,set,remove}xattr calls correspond to
the "client" direction. The client request contains the name
of the extended attribute it wishes to interact with. In this
case, the XattrMap type supplies a `map_client_xattr` method.

Only listxattr is considered a server direction request. In this
case, the xattr map rules can be used to modify or conceal xattrs
that would otherwise be included in the list. For this case,
the XattrMap type supplies a `map_server_xattrlist` method.

Signed-off-by: Connor Kuehl <ckuehl@redhat.com>