Drop mapped "security.capability" on write
Linux will drop the extended attribute "security.capability" on write;
however, as far as the guest is concerned it's possible that they've
remapped "security.capability" to something else. If that's the case,
the host kernel doesn't know to drop the capability automatically.

In these cases, the daemon will uphold this behavior on behalf of the
host kernel.

Relevant virtiofsd upstream commit in mainline QEMU:
e586edcb410543768ef009eaa22a2d9dd4a53846

Signed-off-by: Connor Kuehl <ckuehl@redhat.com>