Set up uid/gid mappings in user_namespace sandbox
When using the namespace sandbox mode with a non-privileged user, set
up 1-to-1 mappings for the current process uid and gid. This allows
both root and a user with the same uid/gid in the VM to access the
files served by us. This scenario fits nicely with the "shared folder"
use case.

Signed-off-by: Sergio Lopez <slp@redhat.com>