stat(): Accept path argument
Both of our stat variants (`fstatat64()` and `statx()`) accept a base FD
and a path relative to it.  In a future where we want to keep the number
of open FDs at a minimum, we are going to need the ability to pass a
non-empty path (so we do not need to open the exact file we want to
stat).

An example is `PassthroughFs.do_lookup`.  Ideally, we do not want to
open the inode, so for stat-ing, we are going to need to
`stat(&p.file, name)`.

This example also illustrates why this patch makes the new `path`
argument a `CStr`: All callers that will want to pass a path only have
`CStr`s available.

On the other hand, it is cumbersome to create a new `CStr` object, so
for simplicity's sake, we take an `Option<CStr>`.  Passing `None` is
translated to an empty `CStr`, so callers do not have to construct it
themselves.  (This translation could hypothetically be done centrally by
`PassthroughFs.stat()`, but instead, this patch does it in both
`stat64()` and `statx()`.  We will want to call `statx()` with an empty
path as of a future patch in this series, so this fits our needs
better.)

Signed-off-by: Max Reitz <mreitz@redhat.com>