Add InodeData implementation to access .file
We want to make storing a `File` in `InodeData` optional, and instead
store a file handle.  For this, we need `InodeData` methods that allow
users to get an open file from `InodeData`, regardless of whether there
is a `File` or just a handle.

There are two methods we need:
- `get_file()`, which returns an `O_PATH` FD,
- `open_file()`, which returns an FD opened with the given flags.

Both functions return an `InodeFile` object, which encapsulates a
`File` object or reference.  We need this because `get_file()` should
not duplicate the existing `File` if there is any, and instead we just
want to return a reference to it.  But if there is only a file handle
(and not a `File`), we need to open it, thus creating a new `File`
object, of which the caller must take ownership (to generally release it
when it goes out of scope).

`InodeFile` must hide this dichotomy.  It does so by providing two
methods to access the file, namely:
- `as_raw_fd()`: Which returns its FD,
- `into_file()`: Which returns an owned `File`.  (An encapsulated
                 reference must be cloned then.)

Note that with the possibility of `InodeFile` owning the `File`,
dropping an `InodeFile` object may mean dropping the `File`, so file
descriptors received with `InodeFile.as_raw_fd()` are only valid for as
long as the `InodeFile` exists (as is the case with any `AsRawFd`
implementation).  Therefore, in places that used to use
`inode_data.file.as_raw_fd()`, we now need a new variable to keep the
`InodeFile` referenced.

After this patch, the only places outside of `InodeData.get_file()` and
`InodeData.open_file()` that access `InodeData.file` are the two
`InodeData` initializers.

Signed-off-by: Max Reitz <mreitz@redhat.com>