Create file handles
Now that we can work with file handles as alternatives to `O_PATH`
`File` objects, we can actually let the functions that create
`InodeData` objects try to generate such handles (before falling back to
`O_PATH` FDs).

The function we use for this (`FileHandle::from_name_at_with_mount_fds`)
checks `mount_fds` to see whether the handle's mount ID is already
mapped there, and if not, it will open the file for which it was
supposed to generate a handle, and put it into `mount_fds` to represent
this mount ID.  (Because this file must be in that mount.)

Generating handles is guarded by the `inode_file_handles` config option,
which is enabled with the `--inode-file-handles` command line switch.

Signed-off-by: Max Reitz <mreitz@redhat.com>