Drop security.capability in setattr() branches
If the security.capability xattr is mapped to something else for the
guest, drop it when executing the uid/gid/trunc branches for setattr.

It must be dropped on all writes and the host kernel won't know to look
for the mapped attribute; only the server knows what the mapped attribute
name is.

Signed-off-by: Connor Kuehl <ckuehl@redhat.com>