impl FromStr for SandboxMode
This commit derives EnumString for SandboxMode
to get FromStr impl.
FromStr impl is needed by the next commit which
replaces clap by structopt.
The attribute parse(try_from_str) in #[derive(StructOpt)]
will generate code to convert cli arg (which is of type &str)
to the expected type if the type impls FromStr. This reduces code.

Signed-off-by: Mihir Luthra <luthramihir708@gmail.com>