Avoid using libc::statx()
`libc::statx()` requires a `statx()` wrapper to be present in the
system's libc, which may or may not be the case.  For example, in glibc
before 2.28, it was not present.

Directly invoking the syscall instead has no real drawbacks (other than
being ugly), because:
- The libc's `statx()` is just a syscall wrapper anyway,
- `libc::syscall()` is always[1] present on Linux systems,
- `libc::SYS_statx` is always[1] present on Linux systems in the gnu and
  musl environments.

`libc::statx()` is only provided on Linux in the gnu environment, so
this does not shrink our coverage.

[1] "Always" means that we can freely choose the rust-libc version
    through Cargo.toml, so we can just choose a version where it is
    present.  This is in contrast to the system's libc, whose version we
    cannot influence.

Resolves: #12
Suggested-by: Connor Kuehl <ckuehl@redhat.com>
Signed-off-by: Max Reitz <mreitz@redhat.com>