Avoid double open for setattr(SIZE)
Before this patch, when truncating a file with no file handle given and
the inode is represented by a file handle, `setattr()` will open a
temporary `O_PATH` FD, and then open another `O_RDWR` FD in the
`SetattrValid::SIZE` branch.

That is ineffective.  Instead, we should open an `O_RDWR` FD right from
the start in that case and save us the `O_PATH` FD.

Note that we could simplify this function very much by always opening an
`O_RDWR` FD (unless a handle was passed) and using that throughout the
function (this would save us all the `*at()` function calls), but I
decided to take a conservative approach here.

Signed-off-by: Max Reitz <mreitz@redhat.com>