Fix protocol features
When virtio-fs uses the DAX option, this means we expect the backend to
send some requests to the VMM, and this is why SLAVE_REQ protocol
feature was exposed. But this is not enough since the requests that are
sent by the backend may contain some file descriptors. That's why we
need the additional protocol feature SLAVE_SEND_FD to be exposed.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>