multikey: Allow multiple alternate keys
Our `MultikeyBTreeMap` allows exactly one alternate key per main key.
We will want more than one alternate key, though: Right now, we only
store the combination st_ino/st_dev/mnt_id as an inode's alternate key.
That is not necessarily sufficient to identify some file, though
(because st_ino can be and is reused when an inode is deleted), so we
want to allow using a file handle as an alternate key, too (because that
handle contains a generation ID that allows to discern when an inode ID
has been reused by a newly created entry).

However, it is possible that we cannot create a file handle, so this
should remain an optional additional alternate key, and not replace the
current st_ino/st_dev/mnt_id combination.  So we need multiple alternate
keys.

This change relaxes some guarantees `MultikeyBTreeMap` used to give us,
mainly the fact that the alternate key would really act like a second
main key: `MultikeyBTreeMap.insert()` was pretty much symmetrical,
regardless of whether the main or the alternate key was changed, only
when determining the the return value did the main key take precedence.

Now, `MultikeyBTreeMap.insert()` only takes the main key.  Alternate
keys have to be added with `MultikeyBTreeMap.insert_alt()` by
associating them with a primary key.

For our use case, that is completely fine.

This change does require some updates to the `MultikeyBTreeMap` tests,
though.

Signed-off-by: Max Reitz <mreitz@redhat.com>