InodeData: Replace alt key by dev/mnt_id
We do not really need to store the alt key here, but only the device and
mount IDs to detect crossmounts.  So store them instead.

(We want this change, because a future patch will make `InodeAltKey`
into an enum that can hold either the current three IDs, or a file
handle.  We would have to always store the former variant, and when
extracting the IDs, we would have to assert that `InodeData.key` is of
that variant, which would be cumbersome.  So it makes more sense to
store what we actually want.)

Signed-off-by: Max Reitz <mreitz@redhat.com>