Add file handles as alternate inode keys
Currently, `InodeData.file_or_handle` is always a `File`.  Therefore,
when an inode is unlinked, its inode ID will remain in use until we drop
the inode from our `PassthroughFs.inodes` map (and thus close the file).
Therefore, `do_lookup()` can safely use the inode ID as a key into that
map (see `InodeAltKey::Ids`), because any inode with an inode ID we find
in `.inodes` (and on the same filesystem) must be the exact same file.

This will change when `.file_or_handle` can store a file handle instead
of a file.  Then, unlinking a file will immediately remove the inode, so
its ID can then be reused by newly created files, even while we still
have the `InodeData` object in our `.inodes` map[1].

So creating a new file can then reuse the old file's inode ID, and
looking up the new file would lead to us finding the old file's
`InodeData`, which is not ideal.

Luckily, just as file handles cause this problem, they also solve it:  A
file handle contains a generation ID, which changes when an inode ID is
reused, so the new file can be distinguished from the old one.  So all
we need to do is use the file handle as the `.inodes` alternate key
instead of the st_ino/st_dev/mnt_id combination (`InodeAltKey::Ids`) we
currently use.

Unfortunately, we cannot rely on being able to generate file handles
every time.  Therefore, every entry should still get an
`InodeAltKey::Ids` alt key, and optionally an additional
`InodeAltKey::Handle` one.  The latter one takes precedence, the former
one is the fallback.

Note that we do not generate `FileOrHandle::Handle()` objects yet, and
so we also do not generate `InodeAltKey::Handle()` alt keys yet.  We
might manually create file handles with some code that is immediately
removed by the next patch again, but that would break the `.filter()`
call in `do_lookup()` that assumes that every inode with a handle alt
key must also have a file handle in `.file_or_handle`.  So we leave
actually creating handle alt keys for the next patch.

[1] If some application in the guest still has the file open, there is
going to be a corresponding `HandleData` object, which holds a `File`.
In such a case, the inode will only go away once every application in
the guest has closed it.  The problem described only applies to cases
where the guest does not have the file open, and it is just in the
dentry cache, basically.

Signed-off-by: Max Reitz <mreitz@redhat.com>