Limit `set_creds()` scope to file modifications
Functions that invoke `set_creds()` generally do so at their beginning,
so the effective UID/GID is changed during their whole scope.  That is
not necessary, we only need the context's eUID/eGID when invoking
functions (i.e. mostly syscalls) that create new nodes, or when writing
to a file.  For example, a `do_lookup()` call after such a syscall does
not need to be done with the context's credentials.

This will become important when we start allowing file handles instead
of `O_PATH` files in `InodeData`: Doing a lookup may entail a
`open_by_handle_at()` syscall, which only root is allowed to do.
(Similarly, even accesses such as `InodeData.file.as_raw_fd()` may
require an `open_by_handle_at()`, so they too will need to be done
outside of the `set_creds()` scope.)

(From a file handle perspective, changing the `set_creds()` scope in
`write()` is unnecessary, because that function works with a
`HandleData` object, which will still always contain a `File`.  However,
it seems sensible to be consistent with how `set_creds()` is used.

Also, we do not really need to create a new block in `write()`, but do
it anyway to make our intent to limit the scope explicit.)

Reported-by: Sergio Lopez <slp@redhat.com>
Signed-off-by: Max Reitz <mreitz@redhat.com>