Do not open special files for `mount_fds`
We must not open special files (i.e. anything but regular files and
directories) with any flag but `O_PATH`.  Right now,
`from_name_at_with_mount_fds()` opens files indiscriminately of their
type, and so this is wrong.

Instead, open the file with `O_PATH` first and use the information we
get from `statx()` (which we call anyway) to determine whether it is
safe to reopen this file with `O_RDONLY`.

Reported-by: Dr. David Alan Gilbert <dgilbert@redhat.com>
Fixes: a80c3be3c59c9c2aaae3a90ebef2b72d151a2e4c
       ("Create file handles")
Signed-off-by: Max Reitz <mreitz@redhat.com>