Update to vhost 0.2
Update to use "vhost 0.2". In addition to pointing to the published
crate instead of the git repository, this change brings us the ability
to listening on existing file descriptors, enabling us to implement
the equivalent of virtiofsd "--fd" argument.

Also update all the required dependencies (vhost-user-backend,
vm-memory, vm-virtio/virtio-queue and vmm-sys-util) and apply the
minimum set of changes to support them.

Ideally, these changes would have been presented as different commits,
but we need to do a single one to avoid breaking bisect.

Signed-off-by: Sergio Lopez <slp@redhat.com>