Add socket-path, deprecate socket, drop sock
Clean up the mess we have with the command line argument to specify
the location of the socket path, and settle on "socket-path" in
accordance with the vhost-user specification.

Here we add "socket-path", deprecate "socket" (allowing it but
printing a warning that's deprecated) and drop "sock" (which we've
been warning it was deprecated for a while).

Signed-off-by: Sergio Lopez <slp@redhat.com>