Clarify and fix inode lookup fallback condition
Looking up an inode is done primarily by its file handle.  When that
fails, we fall back to looking it up by its inode ID.

The explanation for when we cannot fall back was a bit confusing.  Be
more clear about it: We can only fall back if we have an FD open for the
inode we found, because if we do not, then the inode might have been
deleted at some point and its inode ID reused.  Only if we have an FD
open can we be certain that the inode ID remains tied to this respective
inode.

This also means inverting the condition from "InodeData must not contain
a file handle" to "InodeData must contain an FD".

Also, the `handle_altkey.is_none()` part of the condition was wrong, so
drop it.  The idea behind it was that we have no choice to fall back if
we failed to generate a file handle, and so must use the fall back
result no matter what; but that is wrong, we must not use that result
unless we have an FD open.

Signed-off-by: Hanna Reitz <hreitz@redhat.com>