Add safe openat() wrapper
We have a couple of places where we invoke `libc::openat()`, each with
the same repeated explanation why it is safe.  Add an `openat()` wrapper
that converts a `&str` into a `CString` (so we need fewer static C
strings) so that we can make some `libc::openat()` calls simpler.

(We could go even further by also providing a version that takes a
`&CStr` instead of a `&str`, but this would not gain much.)

Signed-off-by: Hanna Reitz <hreitz@redhat.com>