file_handle: Create OpenableFileHandle struct
Right now, creating a `FileHandle` means invoking
`FileHandle::from_name_at_with_mount_fds()`.  This function creates the
file handle, and then ensures that we have a mount FD for it.

The problem with this approach is that creating a file handle fails
altogether when creating the mount FD fails.  However, we do not always
need openable file handles; we also use them as a key for lookups, and
in such cases we do not care whether there is an associated mount FD.
We thus want to separate opening the mount FD from generating the file
handle.

(Parenthesis: When we have an inode that has no associated FD but only a
file handle, this file handle must be openable, i.e. a mount FD for it
must exist.  When we generate a file handle to look up such an inode, it
will thus automatically have a valid associated mount FD.  So in
practice, this is not an issue -- generating file handles for lookups
that must be done by file handle will not fail because opening a mount
FD failed.  However, evidently, this is not a trivial and obvious fact.
So it is better not having to think about this and just to separate
generating the pure file handle from ensuring there is an associated
mount FD.)

This separation is done by introducing a new struct
`OpenableFileHandle`, which is created from a `FileHandle` reference by
invoking `.to_openable()`.  It contains the bulk of the previous
`.from_name_at_with_mount_fds()` function.

In passthrough/mod.rs, we now use `FileHandle` objects for lookups (as
an alternate key), and `OpenableFileHandle` objects in the
`FileOrHandle` enum.

Besides the separation of possible failures during file handle
generation and mount FD opening, there is another reason why we need
this new struct: We want to add refcounting to mount FDs, and for this
to work nicely, `OpenableFileHandle` will need to implement the `Drop`
trait.  We cannot do this for `FileHandle`, as that struct implements
`Copy`, which conflicts with `Drop`.
(Also, to implement `Drop`, we are going to need to add a reference to
the mount FDs map to `OpenableFileHandle`, which would be really ugly in
a structure that is used as a hash map key (i.e., `FileHandle` as part
of `InodeAltKey`).)

Signed-off-by: Hanna Reitz <hreitz@redhat.com>