file_handle: Add OpenableFileHandle.mount_fd_map
Wrap `MountFds.map` in an `Arc`, so we can put a reference to it in
every `OpenableFileHandle` object we create.  We will later use this to
decrement mount FD's refcounts in `OpenableFileHandle`'s `Drop`
implementation.

In this patch, we only make use of a nice side effect: By having a
reference to the mount FD map in `OpenableFileHandle`, we can drop the
corresponding parameter from `OpenableFileHandle.open_with_mount_fds()`
(and accordingly rename it to just `.open()`).

Signed-off-by: Hanna Reitz <hreitz@redhat.com>