file_handle: Add refcounting to MountFd
Once all inodes that contain file handles for a specific mount are
dropped,  we should also drop the respective `MountFd`.  In theory, this
would allow things like receiving a notification when a mount root is
deleted (because this is the node we open to get a mount FD).  In
practice, deleting a mount root is impossible, however.

There may be other cases where dropping the mount FDs is beneficial
besides deletion notifications, though.  For example, this means that
when a guest tells the host to forget about a whole submount, the host
is then free to unmount it.  It could not do that if there were still a
mount FD open for it.

(From my experience, forgetting about a whole submount means using
`--announce-submounts`, then unmounting it in the guest, and then
unmounting also the parent mount, or the mount point is not forgotten.
This may sound cumbersome, but it means that unmounting the whole
virtio-fs filesystem will allow the host to unmount any submount in the
shared directory, which is not possible without mount FD refcounting.)

With refcounting, we need to have `.to_openable()` double-check whether
there is an existing entry in the mount FDs map, namely we need to add a
check right before adding a new entry.  It is possible that a concurrent
thread entered a new entry while we opened a mount FD.  Until now, this
was not a problem, because our new entry would just supersede the old
one -- now, however, that would mean losing the existing refcounts,
which would be bad.

Signed-off-by: Hanna Reitz <hreitz@redhat.com>