Open /proc/self/mountinfo
Generating a file handle yields a mount ID for which it is valid.
Opening the handle however requires specifying an FD on that mount.  To
get this FD, right now we just open the first file that is looked up on
said mount; but we really want to use the mount point instead[1],
because with inotify support coming at some point, we want the inode
belonging to the mount FD to be deletable (such that the guest receives
a notification about the deletion).  There are two reasons why that
means we want to use mount points to get mount FDs:

(1) The mount point can only be deleted once everything else in the
    mount has been deleted, too.  That means all of our references to
    such inodes must be gone, and all file handles requiring the
    respective mount FD are gone, too, so we can close the mount FD at
    that point.
    (To do that, we need mount FD refcounting, which is added by a later
    patch.)

(2) (Host) mount points cannot be deleted anyway.  Attempting to do so
    will return EBUSY.  Using them for mount FDs means that when it
    comes to deletion notifications they do not play a role.

This commit makes virtiofsd-rs open /proc/self/mountinfo, first, if
necessary, in the sandboxing code, and then in `PassthroughFs::new()`
otherwise.  This is just how it works for /proc/self/fd.

We do not make use of mountinfo's contents yet, though; that is left for
the next commit.  Until then, `PassthroughFs::new()`'s `mountinfo_fd`
has to remain unused.

[1] In practice, both are probably going to be the same (the mount point
is probably the first inode of a submount that we look up).  However,
making this explicit is better.

Signed-off-by: Hanna Reitz <hreitz@redhat.com>