Avoid stat() race condition in lookup
When `do_lookup()` can generate a file handle, it will not open an
`O_PATH` FD.  To get stat information, we later stat based on `p_file`
(the parent directory) and `name` (the directory entry name to look up).
However, this can get racy, because the directory entry could be
replaced between generating the file handle and the `stat()` call.

To get rid of the race, unconditionally open an `O_PATH` FD first, and
then get the file handle and stat information from it.

Do the same in `init()`, because it is basically the same code.

Reported-by: Vivek Goyal <vgoyal@redhat.com>
Signed-off-by: Hanna Reitz <hreitz@redhat.com>