file_handle: Add `FileHandle::from_fd()`
This is a shortcut for you want to convert a `File` directly to a
`FileHandle` without the need to specify a relative path.

Coincidentally, this is exactly what we need in passthrough/mod.rs.

Unfortunately, we cannot use this to implement `TryFrom<&File>` because
the result is double-wrapped in `io::Result<>` and `Option<>`.

Signed-off-by: Hanna Reitz <hreitz@redhat.com>