config: Store /proc/self/fd as a File
During configuration, we open /proc/self/fd as a `RawFd` and pass it
around this way.  The clearest problem with this is that we assume this
FD is easily copyable by simply copying its integer value.  However, in
reality, care must be taken to decide whether you really want to use the
same FD in both copies (by copying its integer value) or whether you
want to duplicate it to a new FD (with `dup()`).  It seems especially
dangerous to turn an FD into a `File` object while the FD is still part
of other structures (namely `Sandbox`).

Luckily, we in fact never need to copy the FD.  To make that clear, turn
the FD into a `File` right from the start, and move it around to where
we need it (with `take()`) rather than duplicating it.

This means `passthrough::Config` can no longer derive the `Clone` trait,
but we do not need it anyway (again by using `take()` where
appropriate).

Signed-off-by: Hanna Reitz <hreitz@redhat.com>