file_handle: Pass some errors to the guest
The most common error that can occur when generating a file handle is
`EOPNOTSUPP`, meaning the filesystem does not support file handles.  In
such a case, we should fall back to `O_PATH` FDs, as we already do.

However, in very rare cases it is imaginable that `name_to_handle_at()`
returns other, intermittent errors, like `ENOMEM`.  In such cases it may
not be ideal to fall back to an `O_PATH` FD, or even to proceed with the
inode lookup at all.

As for why, here is an example:

Say we successfully generate a file handle for a new inode, and put it
into its `InodeData`.  When looking it up later, we fail to generate a
file handle with an intermittent failure like `ENOMEM` (`EOPNOTSUPP` is
not an intermittent failure, meaning we would have seen it on our first
`name_to_handle_at()` call already).  So for the lookup, we fall back to
using the inode ID; with this, we find the existing inode, but we cannot
return it, because the inode ID alone cannot identify an inode for which
we have no FD open.  Therefore, we are forced to create a second
`InodeData` for the same inode, which is less than ideal.

That is why we want to abort the lookup once we get an intermittent
failure from `name_to_handle_at()`.  Such failures should be exceedingly
rare (as rare as `ENOMEM` from the kernel can be), so this should not
cause a tangible regression.

Note that when generating an `OpenableFileHandle` from the `FileHandle`,
we continue to ignore all errors (which can arise from trying to open a
mount FD).  That is because the `OpenableFileHandle` plays no part in
the lookup process, and really is optional: If we get one, we can put it
into the `InodeData`; if we do not, we will put the `O_PATH` FD there.

Reported-by: Vivek Goyal <vgoyal@redhat.com>
Signed-off-by: Hanna Reitz <hreitz@redhat.com>