rustfmt: Set imports_granularity to Module
We do not have a consistent style that we use for import statements,
which is a bit of a pain because that is nothing one should have to
think about.

Let `cargo fmt` do the job for us by saying so in rustfmt.toml.  The
"Module" granularity seems to fit our current "style" most closely, so
let's use that one.

Fixes: #16
Signed-off-by: Hanna Reitz <hreitz@redhat.com>