Replace crate "caps" with "capng"
The "caps" crate is elegant, but it doesn't support atomic drop+raise
operations and its multithread support is unclear.

This change replaces it with "capng", which provides a thin layer on
top of libcap-ng. It's not idiomatic, but fits nicely our needs and
it has already been tested on our use cases, as it's the library used
by virtiofsd (the implementation in C).

Signed-off-by: Sergio Lopez <slp@redhat.com>