Drop capabilities before starting the service
As a security measure, drop the process capabilities before starting
the actual file system service, retaining just the ones we really
need (the list was taken from the C implementation).

In case we're spawning a child to run the service (sandbox != none),
clear all the capabilities for the parent, as it only needs to wait
for the child to exit.

Signed-off-by: Sergio Lopez <slp@redhat.com>