Implement KILLPRIV_V2 support
KILLPRIV_V2 is a new FUSE flag that indicates that the filesystem is
responsible for clearing security.capability xattr and clearing setuid
and setgid bits, following these rules:

- clear "security.capability" on write, truncate and chown unconditionally
- clear suid/sgid if following is true. Note, sgid is cleared only if
  group executable bit is set.
   o setattr has FATTR_SIZE and FATTR_KILL_SUIDGID set.
   o setattr has FATTR_UID or FATTR_GID
   o open has O_TRUNC and FUSE_OPEN_KILL_SUIDGID
   o create has O_TRUNC and FUSE_OPEN_KILL_SUIDGID flag set.

We were already clearing "security.capability" on those actions. This
patch takes care of the rest (imitating its implementation in the C
version) and starts announcing KILLPRIV_V2 support to the client (guest
kernel).

Signed-off-by: Sergio Lopez <slp@redhat.com>