Replace Firecraker's seccomp with seccomp-sys
Some distributions require all crate dependencies to be packaged
individually. This implies that depending on the whole Firecracker
repository would make things very hard for us to get virtiofsd-rs
packaged.

This change replaces the dependency on Firecracker's seccomp with
seccomp-sys, which is a crate providing raw bindings for
libseccomp. This means we add the latter as a both a build time and
run time dependency, but that should be fine as the C implementation
also depends on it.

I've also considered adding syscallz, which provides safe wrappers on
top of syscomp-sys, but that one also depends on strut, which we need
to drop from our dependencies due to its problematic versioning
policy.

Signed-off-by: Sergio Lopez <slp@redhat.com>