Update virtio-queue dependency
Update virtio-queue dependency to rev cc1fa35, which brings support
for big-endian machines. Also update vhost-user-backend to rev 264d46d
and vm-memory to v0.7, as we need to keep them in sync with
virtio-queue.

This change also requires us to replace epoll::EventFd with
vmm_sys_util::eventfd, which is nice because is one less crate we need
to depend on.

Signed-off-by: Sergio Lopez <slp@redhat.com>