Set "log-level" default value to "info"
In the previous commit we've added some messages with "info" level
which are useful to known the current status of the server. Set now
the default log level to "info", so those messages will appear unless
the user explicitly chooses a higher log level.

This change is specially important given that we expect virtiofsd to
be usually launched by a manager, such as libvirt, and we want to make
sure to show something about its status in the log files.

Signed-off-by: Sergio Lopez <slp@redhat.com>