Give --inode-file-handles an obligatory argument
We do not want --inode-file-handles to default to a mode where fallbacks
to O_PATH FDs occur: Users may expect that if they ask for file handles,
they always get file handles, and so the current behavior may be
surprising.

We will add exactly such a mode later (where no fallbacks occur), but
for now we only add an argument to --inode-file-handles that lets users
decide which mode to use.  The only current mode ("fallback") is an
obligatory argument to that option, which means that this is an
incompatible change.  It is necessary for 1.0, though, so we do not
cement the "fallback" mode as the default for it.

Note that we need require_equals, or structopt will allow a plain
--inode-file-handles to be shorthand for --inode-file-handles=never,
which is definitely what we do not want.

Signed-off-by: Hanna Reitz <hreitz@redhat.com>