sandbox: Canonicalize shared_dir path
There is no real downside to canonicalizing this path (i.e. to get the
realpath), because by far the likeliest errors to happen are things that
would appear anyway once we try to use the path, like ENOENT.

We need this canonicalization for file handles on sandbox=chroot:
/proc/self/mountinfo's content is not affected by the chroot(), so to
open the mount points listed within, we need to remove the prefix that
is the shared directory path from them.  As far as I can tell, the mount
points listed there are canonicalized, so the shared directory path
needs to be canonicalized, too.  (This stripping is done in a later
patch, here we only canonicalize shared_dir.)

Signed-off-by: Hanna Reitz <hreitz@redhat.com>