Clean up code paths around FileHandle::to_openable
Since we return all errors from `FileHandle::to_openable()` to the
guest, we know for sure that we have an openable handle if and only if
we have any file handle.  Therefore, we no longer need a temporary
`openable_handle` variable that used to potentially be `None` while
`handle` was `Some(_)`, which saves us some lines of code.

Signed-off-by: Hanna Reitz <hreitz@redhat.com>