Add PassthroughFs::get_file_handle_opt()
When generating a file handle, we need to take the `inode_file_handles`
mode into account: In `Never` mode, we skip generating a handle
altogether, in `Fallback` mode, some errors lead to falling back to an
`O_PATH` FD, and in `Mandatory` mode, all errors are fatal.

Centralize this distinction in a single function that can be used by
both places where we generate file handles (`PassthroughFs::do_lookup()`
and `PassthroughFs::init()`).

In a future commit, this will become even more useful for logging errors
or warnings whenever a filesystem does not support file handles.  We
definitely do not want to duplicate that code.

Signed-off-by: Hanna Reitz <hreitz@redhat.com>