mount_fd: Add MPRError type
We want to log file-handle-related errors that indicate a potential
configuration error on the user's end, i.e.:
- When the filesystem in question does not support file handles
  (`FileHandle::from_fd()` returns `Ok(None)`)
- When we have trouble opening a mount FD for a given mount ID
  (`FileHandle::to_openable()` returns `Err(_)`)

In these cases, we want to give the user some more detail than just
`strerror(EOPNOTSUPP)` or `strerror(EINVAL)`.  That is easy in the first
case (just print "FS does not support file handles"), but not so much in
the second case, because only `FileHandle::to_openable()` knows exactly
what has gone wrong.  Therefore, this commit introduces a new error type
`MPRError`, which encapsulates both an `io::Error` and a description
string.  A later commit will have `FileHandle::to_openable()` return
this error type and set the description as appropriate.

There is more, though.  Because these logged error messages are intended
to warn the user that their configuration might need a fix, we only want
to log them once per filesystem, to prevent the log from being spammed.
To implement this, we add a `MountFds::error_for()` method that returns
an `MPRError` for a given mount ID.  This method checks whether we have
already generated an error for this ID, and if so, marks the error as
silent so that it will not be logged.

This `MountFds::error_for()` method also adds the mount ID and (if
possible) the corresponding mount point path to the `MPRError` object so
that the description, when logged, can be prefixed with this
information.

Finally, note that with these properties (only logged once per
filesystem, and shows which filesystem the error has appeared on) this
error object will not only be useful for `FileHandle::to_openable()`,
but also for the other case, when a filesystem does not support file
handles.  `PassthroughFs::get_file_handle_opt()` (as the caller of
`FileHandle::from_fd()`) will (as of a later commit) construct an
`MPRError` object itself in order to log warnings/errors for when file
handles are not supported.

Signed-off-by: Hanna Reitz <hreitz@redhat.com>